Adding group exec
Create a zero-value run.Group, and then add actions to it.
Actions are defined as a pair of functions: an execute function,
which should run synchronously; and an interrupt function, which,
when invoked, should cause the execute function to return.

Finally, invoke Run, which concurrently runs all of the action,
waits until the first action exits, invokes the interrupt functions,
and finally returns control to the caller only once all action have
returned.

This general-purpose API allows callers to model pretty much any
runnable task, and achieve well-defined lifecycle semantics for the
group.