Backoff schedule
In order to make the task library more powerful in terms of scheduling,
the task function now allows to carry forward a returned error. Using
this error allows us to identify if a backoff is required during a
schedule. With this setup we can then do clever things around burst
timings for tasks and sleeping patterns.