Self terminate
The following changes allows a task to self terminate. If a task has no
way forward and is blocked, it makes no sense to continue the task
looping. Instead the task can state that it want's to terminate and if
that's the case we should allow it to do so.