Allow occupy to skip tasks
Sometimes we don't have a task to execute on and it would be wrong to
expect occupy to do something. Instead we add the concept of skipping
which will keep the abstraction inside of occupy. This means that occupy
can identify when a skip is called and just continue onwards.