Tomb keep-alive
There are scenarios where we want to have a tomb keep running until it's
killed, rather than it self terminating once the go routines have
finished.

This is somewhat pushing the limits of a tomb and we could have another
name for this, but essentially we are where we are.